+++ src/plugin.cpp
@@ -32,4 +32,5 @@ void init(Plugin* p) {
     p->addModel(modelDrums64);
     p->addModel(modelObjects64);
     p->addModel(modelGrains64);
+    p->addModel(modelMicro64);
 }