+++ src/plugin.hpp
@@ -242,3 +242,4 @@ extern Model* modelRhythm64;
 extern Model* modelDrums64;
 extern Model* modelObjects64;
 extern Model* modelGrains64;
+extern Model* modelMicro64;